+++ src/replay/index.ts
@@ -1,9 +1,1 @@
-// Runtime
-export { algebraicToCoords, coordsToSquare, squareToCoords } from '#board/board-coords';
-
-// Types
 export type { Action, BaseAction, CaptureAction, MoveAction, PromoteAction } from '#types/actions';
-export type { BoardCoord, Square } from '#board/board-coords';
-export type { BoardPieceName, Piece, PromotedPieceName } from '#board/piece-names';
-export type { PlayerColor } from '#types/tokens';
-export type { ReplayMode } from '#replay/replay-mode';